refactor: 픽스쳐 정리
