+++ mvc/src/test/java/nextstep/mvc/controller/tobe/HandlerAdapterRegistryTest.java
@@ -3,12 +3,14 @@ package nextstep.mvc.controller.tobe;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.NoSuchElementException;
-import nextstep.mvc.controller.tobe.fixture.AnnotationController;
-import nextstep.mvc.controller.tobe.fixture.FakeHandlerAdapter;
-import nextstep.mvc.controller.tobe.fixture.ImplementedController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import samples.FakeHandlerAdapter;
+import samples.ImplementedController;
+import samples.TestController;
 
 class HandlerAdapterRegistryTest {
 
@@ -32,8 +34,10 @@ class HandlerAdapterRegistryTest {
         registry.add(new ControllerHandlerAdapter());
         registry.add(new HandlerExecutionHandlerAdapter());
 
-        final var controller = new AnnotationController();
-        final var execution = new HandlerExecution(controller, AnnotationController.class.getDeclaredMethod("get"));
+        final var controller = new TestController();
+        final var execution = new HandlerExecution(controller,
+                TestController.class.getDeclaredMethod("findUserId", HttpServletRequest.class,
+                        HttpServletResponse.class));
         final var handlerAdapter = registry.getHandlerAdapter(execution);
 
         assertThat(handlerAdapter).isInstanceOf(HandlerExecutionHandlerAdapter.class);
