+++ mvc/src/test/java/nextstep/mvc/controller/tobe/HandlerMappingRegistryTest.java
@@ -7,7 +7,6 @@ import static org.mockito.Mockito.when;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.NoSuchElementException;
-import nextstep.mvc.controller.tobe.fixture.AnnotationController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -17,11 +16,11 @@ class HandlerMappingRegistryTest {
     @Test
     void getHandler() {
         final var registry = new HandlerMappingRegistry();
-        final var handlerMapping = new AnnotationHandlerMapping("nextstep.mvc");
+        final var handlerMapping = new AnnotationHandlerMapping("samples");
         registry.add(handlerMapping);
 
         final var request = mock(HttpServletRequest.class);
-        when(request.getRequestURI()).thenReturn(AnnotationController.url);
+        when(request.getRequestURI()).thenReturn("/get-test");
         when(request.getMethod()).thenReturn("GET");
 
         assertDoesNotThrow(() -> registry.getHandler(request));
