--- mvc/src/test/java/nextstep/mvc/controller/tobe/fixture/OtherAnnotationController.java
@@ -1,7 +1,0 @@
-package nextstep.mvc.controller.tobe.fixture;
-
-import nextstep.web.annotation.Controller;
-
-@Controller
-public class OtherAnnotationController {
-}
