+++ mvc/src/test/java/samples/FakeHandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe.fixture;
+package samples;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
