+++ mvc/src/test/java/samples/OtherTestController.java
@@ -1,0 +1,7 @@
+package samples;
+
+import nextstep.web.annotation.Controller;
+
+@Controller
+public class OtherTestController {
+}
